brainstone/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bluepy==1.3.0
certifi==2023.5.7
cffi==1.15.1
cryptography==43.0.1
charset-normalizer==3.1.0
idna==3.4
influxdb-client==1.36.1
lywsd02==0.0.9
lywsd03mmc==0.1.0
psycopg2==2.9.6
pycparser==2.21
python-miio==0.5.12
python-dateutil==2.8.2
python-dotenv==1.0.0
randmac==0.1
reactivex==4.0.4
requests==2.31.0
scapy==2.5.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.5.0
typing-extensions==4.6.2
urllib3==2.0.2
